requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
django-environ==0.9.0
django-filter==22.1
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
Django~=4.1.3
djangorestframework-simplejwt==4.8.0
djangorestframework==3.14.0
djoser==2.1.0
docutils==0.19
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==21.3
pip==22.3
psycopg2-binary==2.9.4
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.4
PyYAML==6.0
requests-oauthlib==1.3.1
requests==2.28.1
ruamel.yaml.clib==0.2.7
ruamel.yaml==0.17.21
setuptools==60.2.0
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.12
wheel==0.37.1
celery==5.2.7
django-celery-results==2.4.0
redis==4.3.4